Pass ScanData to ParserPlugin instead of InputStream
Introduction of ScanData allows multi-file scans to be parsed by plugin
parsers. Both sample-parser and blackduck-plugin were adapted to the new
API.

Introduce SessionStore to keep track of in-process scan sessions.
Session store keeps mapping from ParsingSession to the associated scan
file, that is stored locally util session ends.

Scan cancelation messages and periodic session store cleanup needs to be
introduced to cleanup session data that are not needed anymore (not part
of this commit).

Also improve sample-parser to return scanDate and fill in some
Vulnerability fields that are most visible in SSC UI.